Moved nnls and nanoflann to lib
